refactor: split dashboard app and refresh CLI docs
- Extract dashboard readers, projects, prompts, setup/quality, and terminal flows into focused browser scripts.
- Load and test the split dashboard assets, with Knip coverage for classic script entry points.
- Refresh README and docs command examples to use `npx goat-flow`.
- Clean up dead type imports, formatting, and record the dashboard script split lesson.
