+++ docs/coding-standards/conventions.md
@@ -59,10 +59,10 @@ bash -n scripts/*.sh scripts/maintenance/*.sh                                   
 bash scripts/preflight-checks.sh         # Full preflight gate
 
 # CLI commands (after build)
-goat-flow audit .                        # Validate setup correctness
-goat-flow audit . --harness              # AI harness completeness checks
-goat-flow setup --agent claude           # Generate setup prompt
-goat-flow quality . --agent claude       # Generate quality-assessment prompt
+npx goat-flow audit .                        # Validate setup correctness
+npx goat-flow audit . --harness              # AI harness completeness checks
+npx goat-flow setup --agent claude           # Generate setup prompt
+npx goat-flow quality . --agent claude       # Generate quality-assessment prompt
 ```
 
 ## Conventions
