+++ src/dashboard/index.html
@@ -339,8 +339,14 @@
               x-text="agent.id"
             ></button>
           </template>
+          <template x-for="agent in agentSkeletonList" :key="'skel-' + agent.id">
+            <span
+              class="gf-ab gf-ab-loading"
+              x-text="agent.id"
+            ></span>
+          </template>
           <span
-            x-show="installedAgents.length === 0"
+            x-show="installedAgents.length === 0 && agentsLoaded"
             class="gf-ab"
             style="cursor: default; font-style: italic"
             >no agents</span
@@ -481,6 +487,11 @@
       </main>
     </div>
 
+    <script src="/assets/dashboard-readers.js"></script>
+    <script src="/assets/dashboard-setup-quality.js"></script>
+    <script src="/assets/dashboard-projects.js"></script>
+    <script src="/assets/dashboard-prompts.js"></script>
+    <script src="/assets/dashboard-terminal.js"></script>
     <script src="/assets/app.js"></script>
   </body>
 </html>
