+++ src/dashboard/preset-prompts.json
@@ -99,9 +99,9 @@
   },
   {
     "id": "test",
-    "name": "Testing Gap & Overlap Analysis",
-    "desc": "PR-scoped gap & overlap analysis with time-budgeted plan",
-    "prompt": "/goat-qa Analyse the testing gaps in my current PR (gh pr view on the current\nbranch; if no PR, use commits between HEAD and deploy).\n\nScope the plan to what I will ACTUALLY run today:\n- modality: [UI | SQL | CLI | mixed - ask if unclear]\n- time budget: [N minutes]\n- environment: [local | mirror | production-verify-only]\n- ticket type: [bug | feature | devops]\n\nExclude what's already covered. For each of these, tell me in ONE LINE\nwhat it proves for THIS PR (so I can sanity-check your assumptions):\n- PHPStan (type/null safety, undefined methods/constants)\n- Symfony + Doctrine validation in CI\n- Access-control scanner in CI (note: known false-negative rate -\n  ~150 flagged routes; not a substitute for per-endpoint checks)\n- Existing PHPUnit tests touching changed symbols (grep tests/). Note:\n  unit test sparsity is by design in this codebase, not a gap.\n- Existing Cypress tests touching changed flows - NAME the test files\n  and assess assertion depth (page-loads-only vs verifies controls/data).\n  Do not assume shallow.\n- Dev's own testing notes on the PR\n- Copilot reviewer comments on the PR (surface, don't duplicate)\n\nThen, before proposing tests, answer these four questions:\n\nA. ENTRY POINTS: list every path that triggers the changed code\n   (portal, API, OLB, cron, webhook). Flag any where behaviour may differ.\nB. PRODUCTION DATA SHAPE: does this touch columns that could be NULL,\n   legacy, or from abandoned workflows in production but clean in base\n   data? Name the columns.\nC. ACCESS CONTROL: does this touch patient/practitioner data or cross\n   PracGroup/LinkPaG boundaries? If yes, list endpoints that need IDOR\n   and tenant-boundary probing.\nD. PRIOR INCIDENTS: search the repo for the changed files/symbols -\n   has this area broken before? Link the tickets.\n\nLabel every claim in A–D as VERIFIED (read from code/config) or\nINFERRED (pattern/naming-based). List what you couldn't determine.\n\nDeliver in this order:\n1. One-line overlap audit (the exclusions above, filled in).\n2. A–D answers, with VERIFIED/INFERRED labels.\n3. 5–10 residual tests in plain English, each with a time estimate\n   summing to my budget. Rank by signal-per-minute, not completeness.\n   For each test, state the specific failure mode it would catch.\n4. Sufficiency verdict: is this budget enough for the risk level? If\n   not, name the single scenario most likely to cause a prod incident\n   and how much time it would add.\n5. One \"action required on PR\" item if testing can't compensate for a\n   code-level issue (deleted guard, missing voter, unbackfilled\n   migration, unresolved reviewer comment).",
+    "name": "Targeted UI Testing Plan",
+    "desc": "Code-change risk analysis with targeted UI testing tasks",
+    "prompt": "/goat-qa Use gh to carefully read the GitHub issue link mentioned below and all comments. Analyse the code changes in the pull request. Then git checkout the branch mentioned. Analyse the code changes locally to determine a highly targeted testing plan.\n\nDo NOT include what the developer has already tested, or what will be covered by static code analysis testing or view-all-pages acceptance testing.\n\nGive me a targeted UI testing plan based on code-change risk, 3-10 tasks around 5-10 minutes each. Present as a markdown table with columns: Test Task, Why Important.\n\nAsk questions if required to achieve the goal of a targeted testing plan that saves time by not testing what is already covered, but still tests based on code-change risk. UI testing only on localdev.",
     "cat": "qa"
   },
   {
