+++ src/dashboard/views/prompts.html
@@ -177,7 +177,11 @@
       </div>
 
       <!-- Right preview -->
-      <div class="gf-card" style="padding: 18px 20px; min-width: 0" x-data="{ extraContext: '' }">
+      <div
+        class="gf-card"
+        style="padding: 18px 20px; min-width: 0"
+        x-data="{ extraContext: '' }"
+      >
         <template x-if="selectedPreset">
           <div>
             <div
